Implement authorization in tasks actions

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateTaskRequest;
 use App\Http\Resources\TaskResource;
 use App\Models\Task;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -16,6 +17,8 @@ class TaskController extends Controller
      */
     public function index()
     {
+        Gate::authorize('viewAny', Task::class);
+        
         return TaskResource::collection(auth()->user()->tasks()->get());
     }
 
@@ -24,6 +27,8 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
+        Gate::authorize('create', Task::class);
+
         $task = $request->user()->tasks()->create($request->validated());
 
         return TaskResource::make($task);
@@ -34,6 +39,8 @@ class TaskController extends Controller
      */
     public function show(Task $task)
     {
+        Gate::authorize('view', $task);
+
         return TaskResource::make($task);
     }
 
@@ -46,7 +53,8 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
-        //
+        Gate::authorize('update', $task);
+
         $task->update($request->validated());
 
         return TaskResource::make($task);
@@ -57,7 +65,8 @@ class TaskController extends Controller
      */
     public function destroy(Task $task)
     {
-        //
+        Gate::authorize('delete', $task);
+
         $task->delete();
 
         return response()->json(['message' => 'Task successfully deleted.'], 200);
